Fix for #1061.
The root problem was colliding editor ids.  The default value of a Workbench editor's id is its name.  If the name is not unique, then we have a problem with certain actions (in particular for #1061, closing the wrong editor panel and then trying to do things with it).

This problem is resolved by using a random uuid as the default id in the MEditor mixin.  The observed problem can still occur if a subclass were to use a non-unique id for editor windows, but this is at least then an application error that they have introduced into their code.