Clarified the purpose of TaskLayout construction args.
TaskLayout's init takes as arguments *items and **traits. At first glance, this makes it seem as if one could initialize TaskLayout.items by calling TaskLayout(items=some_items). However, doing this would create a task layout with items=() and traits={'items': some_items}.

This behavior is misleading.

This commit renames *items to *args and **traits to **kwargs to clarify how the arguments will be interpreted in the constructor.