Skip focus test on Linux.
A widget only has focus True on Linux if it is the active widget, and we
can't (or don't want) to force that to be the case in the test run.
Forcing a widget to be active is very disruptive for developers who may
be running the tests on their working machine.