Move schemas and related code out of Tasks
Schemas are generic, this PR moves the Schema code out of pyface.tasks and
into pyface.action, making it available for use outside of tasks.  Additionally
a version of ActionManagerBuilder is included to provide a Tasks-indpendednt
way of assembling the Schemas into action managers.

Additionally, the code is slightly modified so that it does not import toolkit
code as a side-effect.

The old locations are kept so imports should still work.  They will be removed
at some yet-to-be-determined future point.