Move schemas and related code out of Tasks (#1076)
* Move schemas and related code out of Tasks

Schemas are generic, this PR moves the Schema code out of pyface.tasks and
into pyface.action, making it available for use outside of tasks.  Additionally
a version of ActionManagerBuilder is included to provide a Tasks-indpendednt
way of assembling the Schemas into action managers.

Additionally, the code is slightly modified so that it does not import toolkit
code as a side-effect.

The old locations are kept so imports should still work.  They will be removed
at some yet-to-be-determined future point.

* Remove some imports that end-up importing toolkit-specific code.

* Update pyface/action/schema/tests/test_action_manager_builder.py

Co-authored-by: aaronayres35 <36972686+aaronayres35@users.noreply.github.com>

* Apply suggestions from code review

Co-authored-by: aaronayres35 <36972686+aaronayres35@users.noreply.github.com>

* Fixes based on PR review.

* Push the discovery of additional toolbar schemas to base builder class.

* Clean up imports.

* Update extra_actions trait.

Co-authored-by: aaronayres35 <36972686+aaronayres35@users.noreply.github.com>