Python 3.11 Support (#1210)
This supports Python 3.11 with PySide 6.4 and PyQt 5. PySide 2 doesn't
seem to be readily available for Python 3.11, so we're not testing with
it. All of the changes for this still work with the existing set of
tests for Python 3.8 and Python 3.10 with PySide2, PySide6 and PyQt5 on
all platforms, and also with EDM on Python 3.8

I have manually checked that the examples work, modulo #1212 which is an
orthogonal issue to this.

So this is probably as good as it is going to get - we will likely find
additional issues with the new Enums as we start to see real work done
with it.

As a side-effect, fixes #1202