Workaround for end-of-process segfault on Linux (#1213)
On Linux, calling `QApplication.thread()` appears to cause an
end-of-process segfault. We've been seeing this in the test suites of
several applications using the ETS stack.

I've opened an upstream bug report here:
https://bugreports.qt.io/browse/PYSIDE-2254

In the meantime, we can avoid _most_ calls to `QApplication.thread`
(those not involving a background thread): it's used in a `moveToThread`
operation, but we can skip that operation if the current thread _is_ the
GUI thread.

This PR makes that change, and adds issue references. I suspect we'll
still have the end-of-process segfault problem from code that uses
`GUI.invoke_after` / `GUI.invoke_later` / `GUI.set_trait_later` etc.
from background threads.