Move `pyface.ui.qt4` to `pyface.ui.qt` (#1223)
This moves `pyface.ui.qt4.*` to `pyface.ui.qt` and helps with gradually
deprecating all uses of "qt4". The current state is internal
consistency, optional hooks for backwards compatibility for applications
that depend on Pyface.

Changes for downstream libraries and applications should be fairly
simple: replacing `qt` with `qt4` in appropriate import statements.

But there are some circumstances where this may not be feasible:

- where the end-user doesn't control the code-base (eg. someone who
pip-installs Mayavi and gets the new version of Pyface automatically)
- where an application developer needs to make a newer version of pyface
work with code they don't control that imports from `pyface.ui.qt4.*`
- where a library developer wants to be compatible with newer and older
versions of pyface

This PR solves two of these problems by providing opt-in import hooks
that replace imports of `pyface.ui.qt4.*` with corresponding imports of
`pyface.ui.qt.*` so that `sys.modules` points to the new location
(avoiding issues with duplicated objects and modules which would happen
with some other strategies). These hooks are not available by default,
but can be accessed in a number of ways:

- end-users can set environment variables to turn it on:
`ETS_QT4_IMPORTS` to directly turn on the import hooks, and
`ETS_TOOLKIT=qt4` (as opposed to `ETS_TOOLKIT=qt`) with the assumption
that this is an older environment
- application developers can either set `ETSConfig.toolkit = "qt4"` (as
opposed to `ETSConfig.toolkit = "qt"`) with the assumption that this is
flagging that this is an application made with older assumptions about
import locations; or if they need more control, they can install the
import hooks into `sys.meta_path` in a way that works best for them and
any other import hooks that they might be using.

The third case of a library developer who wants to support newer and
older versions should be done through the usual mechanisms (eg. try
importing `pyface.ui.qt.foo` and if it fails import `pyface.ui.qt4.foo`;
or looking at pyface version numbers). An alternative which will work in
almost all cases is to instead use `toolkit_object()` to get the thing
that they want, which will automatically get it from the right place.

Done:
- [x] check about pickle compatibility with mementos: Tasks and
Workbench mementos are independent of toolkit
- [x] migration documentation (added to toolkits section)

Note:

This ended up taking about a day longer than expected because the
original working version of the code was using a deprecated API that
would be removed in Python 3.12, so it needed re-implementation after
digging in to understand what the new importlib API was doing. In doing
so it has been generalized a bit, since we expect to also do this in
TraitsUI at a minimum.

Fixes #560

---------

Co-authored-by: Mark Dickinson <mdickinson@enthought.com>