Delay imports in api modules if toolkit will be imported (#1233)
This is uses module `__getattr__` and `__dir__` based on [PEP
562](https://peps.python.org/pep-0562/) which is available now we are
dropping Python 3.6 support.

Basic strategy is to handle two classes of imports - those which come
directly via the toolkit object, and those which are via an import
relative to the api module. This allows us to eventually deprecate and
remove stub modules which do nothing more than import using the toolkit,
but we don't do that in this PR.

We test by checking that every public attribute returned by `dir()` can
be accessed.

Includes a few drive-by fixes to avoid toolkit selection where where
possible. The most significant is adding formal `Interface` classes for
`ActionManager` and its subclasses (and using those on
`ApplicationWindow`). It also cleans up a deprecated trait on
`ApplicationWindow`.