Refactor fields (#1234)
This is a refactor of the Field class to split it into display and
editable variants: base `Fields` may not be editable, `EditaleFields`
add editability (ie. the value can be changed by the user). Includes
tests and documentation, including fixes to make the documentation
actually compile.

Adds a LabelField and ImageField to demonstrate the non-editable Fields.

---------

Co-authored-by: Mark Dickinson <mdickinson@enthought.com>