Remove macos-latest from ets-from-source workflow (#1278)
Our `ets-from-source` workflow is currently failing, for the simple and
shallow reason that we're not currently in a position to run EDM-based
tests on macOS/ARM (missing runtime for Python 3.8; missing dependencies
for Python 3.11). This PR removes `macos-latest` from the test matrix,
and makes a couple of other drive-by cleanups.